Bind the parsing error where scala 3 can still be instrumented
The parsing error tests failed for every scala 3 build. Scala 3 compiles the
forwarders a class inherits from a trait as bridge methods, and bridge methods
are not advised, so the advice on the parser's own emit never ran, nothing was
bound to the ErrorInfo and every span fell back to a name of HTTP. Scala 2
compiles the same forwarders as ordinary methods, which is why only the scala 3
suite failed.

Every scala version puts the body of a concrete trait method in a static method
named after it with a $ appended, and the forwarder, the default method and the
trait itself all call that, so it is the one place all the callers meet.

parseMessage and parseRequestTarget are declared by the parser rather than
inherited, so they are advised in both and are left alone.
