+++ instrumentation/pekko/pekko-http-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/pekkohttp/v1_0/server/PekkoHttpServerInstrumentationModule.java
@@ -35,6 +35,7 @@ public class PekkoHttpServerInstrumentationModule extends InstrumentationModule 
         new HttpPrepareAttributesInstrumentation(),
         new HttpServerBluePrintInstrumentation(),
         new GraphInterpreterInstrumentation(),
+        new HttpMessageParserInstrumentation(),
         new HttpRequestParserInstrumentation(),
         new ParsingErrorHandlerInstrumentation(),
         new PekkoHttpServerSourceInstrumentation(),
